added controllers, services, repos, entities and tests

+++ src/main/java/com/backend/tasks/service/org/impl/OrganizationServiceImpl.java
@@ -4,7 +4,6 @@ import com.backend.tasks.entity.Organization;
 import com.backend.tasks.exception.ResourceNotFoundException;
 import com.backend.tasks.repository.OrganizationRepository;
 import com.backend.tasks.service.org.OrganizationService;
-import javassist.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -19,11 +18,11 @@ public class OrganizationServiceImpl implements OrganizationService {
     OrganizationRepository organizationRepository;
 
     @Override
-    public Organization getOrganizationById(Long id) throws NotFoundException {
+    public Organization getOrganizationById(Long id) throws ResourceNotFoundException {
         Optional<Organization> organization = organizationRepository.findById(id);
 
-        if(!organization.isPresent())
-            throw new ResourceNotFoundException();
+        if (!organization.isPresent())
+            throw new ResourceNotFoundException("Organization with ID=" + id + " not found");
 
         return organization.get();
     }
@@ -42,7 +41,13 @@ public class OrganizationServiceImpl implements OrganizationService {
 
     @Override
     public Organization update(Long id, Organization organization) {
-        return organizationRepository.findById(id).map(org -> {
+
+        Optional<Organization> currentOrg = organizationRepository.findById(id);
+
+        if (!currentOrg.isPresent())
+            throw new ResourceNotFoundException("Organization with ID=" + id + " not found");
+
+        return currentOrg.map(org -> {
             org.setName(organization.getName());
             return organizationRepository.save(org);
         }).get();
@@ -50,6 +55,11 @@ public class OrganizationServiceImpl implements OrganizationService {
 
     @Override
     public void delete(Long id) {
+        Optional<Organization> org = organizationRepository.findById(id);
+
+        if (!org.isPresent())
+            throw new ResourceNotFoundException("Organization with ID=" + id + " not found");
+
         organizationRepository.deleteById(id);
     }
 }
